refactor(approve): skip simulateContract
USDT on Ethereum has different ABI of approve function which results in successful gas estimation but simulateContract failed
